+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* refactor hide checkbox to single function
 
 #### 2.0.2 / 2023-08-18
 * add single file plugin to `$plugin_dirnames`