+++ CHANGES.md
@@ -2,6 +2,7 @@
 * override `WP_Plugins_List_Table` to add filter and restructuring of PR
 * add filter to hide plugin row checkbox
 * convert to static class
+* update plugin card description for clarity
 
 #### 2.0.2 / 2023-08-18
 * add single file plugin to `$plugin_dirnames`