+++ CHANGES.md
@@ -4,6 +4,7 @@
 * convert to static class
 * update plugin card description for clarity
 * update Requires WP to 6.4 due to `wp_admin_notice()` use
+* many more updates to coincide with refactoring of PR
 
 #### 2.0.2 / 2023-08-18
 * add single file plugin to `$plugin_dirnames`